Add a dashed border around the about section
This adds a little more separation between the about section and the
rest of the page. Previously, a larger font size was used to achieve
this, but it sometimes felt like it blended with the rest of the text on
the page too easily.

The dashed border also adds a paper-like touch.